pyproject.toml
[project]
name = "socket-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.32",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.7.0",
    "email-validator>=2.1.0.post1",
    "pydantic-settings>=2.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "PyQt6>=6.7.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
server = "server.main:run"
client = "client.main:run"
